Renaming import aliases
