+++ app/[lang]/about/GenreSolarSystem.tsx
@@ -2,9 +2,9 @@
 
 import { useEffect } from "react";
 
-import { useIsDesktop } from "@lib/useMediaQuery";
-import { colors } from "@utils/getColors";
-import { getLuminance, TEXT_LUMINANCE_TRESHOLD } from "@utils/getLuminance";
+import { useIsDesktop } from "~lib/useMediaQuery";
+import { colors } from "~utils/getColors";
+import { getLuminance, TEXT_LUMINANCE_TRESHOLD } from "~utils/getLuminance";
 
 const topGenres = [
 	{
