+++ app/[lang]/login/page.tsx
@@ -1,12 +1,12 @@
 import { getServerSession } from "next-auth";
 import { redirect } from "next/navigation";
 
-import { Container } from "@components/Container";
-import { Section } from "@components/Section";
+import { Container } from "~components/Container";
+import { Section } from "~components/Section";
 import { LoginButton } from "./LoginButton";
 
-import { Locale } from "@/i18n.config";
-import { getDictionary } from "@/utils/getDictionary";
+import { Locale } from "~/i18n.config";
+import { getDictionary } from "~/utils/getDictionary";
 
 export const metadata = {
 	title: "Milovan Gudelj - Login",
