+++ app/[lang]/music-stats/page.tsx
@@ -2,15 +2,15 @@ import Image from "next/image";
 import { headers } from "next/headers";
 import { redirect } from "next/navigation";
 
-import { Artist, Track } from "@lib/types";
-import { getPalette, Palette } from "@utils/getPalette";
+import { Artist, Track } from "~lib/types";
+import { getPalette, Palette } from "~utils/getPalette";
 
-import { Locale } from "@/i18n.config";
-import { getDictionary } from "@/utils/getDictionary";
+import { Locale } from "~/i18n.config";
+import { getDictionary } from "~/utils/getDictionary";
 
-import { BigAssStar } from "@components/BigAssStar";
-import { Container } from "@components/Container";
-import { Section } from "@components/Section";
+import { BigAssStar } from "~components/BigAssStar";
+import { Container } from "~components/Container";
+import { Section } from "~components/Section";
 
 import { FormData } from "./ControlsBar";
 import { StatsSection } from "./StatsSection";
