+++ app/api/auth/[...nextauth]/route.ts
@@ -2,7 +2,7 @@ import NextAuth, { AuthOptions, Account, Session, User } from "next-auth";
 import { AdapterUser } from "next-auth/adapters";
 import { JWT } from "next-auth/jwt";
 import SpotifyProvider from "next-auth/providers/spotify";
-import { spotifyApi, SPOTIFY_AUTH_URL } from "@lib/spotify";
+import { spotifyApi, SPOTIFY_AUTH_URL } from "~lib/spotify";
 
 export interface ExtendedToken extends JWT {
 	accessToken?: Account["access_token"];
