+++ app/api/getUser/route.ts
@@ -1,11 +1,11 @@
 import { NextRequest, NextResponse } from "next/server";
 import { getServerSession } from "next-auth";
 
-import { spotifyApi } from "@lib/spotify";
+import { spotifyApi } from "~lib/spotify";
 import {
 	authOptions,
 	ExtendedSession,
-} from "@/app/api/auth/[...nextauth]/route";
+} from "~/app/api/auth/[...nextauth]/route";
 
 export async function GET(req: NextRequest, res: NextResponse) {
 	const session = (await getServerSession(authOptions)) as ExtendedSession;
