+++ components/NowPlaying/NowPlaying.tsx
@@ -5,9 +5,9 @@ import Image from "next/image";
 import useSWR from "swr";
 import { SpotifyLogo } from "@phosphor-icons/react";
 
-import fetcher from "@lib/fetcher";
-import { NowPlayingSong } from "@lib/types";
-import { useIsomorphicLayoutEffect } from "@utils/useIsomorphicLayoutEffect";
+import fetcher from "~lib/fetcher";
+import { NowPlayingSong } from "~lib/types";
+import { useIsomorphicLayoutEffect } from "~utils/useIsomorphicLayoutEffect";
 
 export const NowPlaying = ({
 	title,
