+++ sanity/schema.ts
@@ -1,6 +1,6 @@
 import { type SchemaTypeDefinition } from 'sanity'
 
-import { project } from "@/sanity/schemas";
+import { project } from "~/sanity/schemas";
 
 export const schema: { types: SchemaTypeDefinition[] } = {
 	types: [project],
