+++ utils/getDictionary.ts
@@ -1,11 +1,11 @@
 import "server-only";
-import type { Locale } from "@/i18n.config";
+import type { Locale } from "~/i18n.config";
 
 // We enumerate all dictionaries here for better linting and typescript support
 // We also get the default import for cleaner types
 const dictionaries = {
-	en: () => import("@/dictionaries/en.json").then((module) => module.default),
-	it: () => import("@/dictionaries/it.json").then((module) => module.default),
+	en: () => import("~/dictionaries/en.json").then((module) => module.default),
+	it: () => import("~/dictionaries/it.json").then((module) => module.default),
 };
 
 export const getDictionary = async (locale: Locale = "en") => {
